Rework oxen_coin lib to avoid exposing native functions
