+++ lib/src/wallet/transaction/pending_transaction.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:oxen_coin/transaction_history.dart' as transaction_history;
-import 'package:oxen_coin/structs/pending_transaction.dart';
+import 'package:oxen_coin/oxen_coin_structs.dart';
 import 'package:oxen_wallet/src/wallet/oxen/oxen_amount_format.dart';
 
 class PendingTransaction {
