+++ lib/src/wallet/transaction/transaction_info.dart
@@ -1,5 +1,5 @@
 import 'package:oxen_wallet/src/wallet/oxen/oxen_amount_format.dart';
-import 'package:oxen_coin/structs/transaction_info_row.dart';
+import 'package:oxen_coin/oxen_coin_structs.dart';
 import 'package:oxen_wallet/src/util/parseBoolFromString.dart';
 import 'package:oxen_wallet/src/wallet/transaction/transaction_direction.dart';
 import 'package:oxen_wallet/src/domain/common/format_amount.dart';
