+++ oxen_coin/lib/oxen_coin_structs.dart
@@ -1,0 +1,7 @@
+library oxen_coin;
+
+export 'src/structs/account_row.dart';
+export 'src/structs/pending_transaction.dart';
+export 'src/structs/stake_row.dart';
+export 'src/structs/subaddress_row.dart';
+export 'src/structs/transaction_info_row.dart';
