+++ oxen_coin/lib/src/exceptions/connection_to_node_exception.dart
