+++ oxen_coin/lib/src/exceptions/creation_transaction_exception.dart
