+++ oxen_coin/lib/src/exceptions/setup_wallet_exception.dart
