+++ oxen_coin/lib/src/exceptions/wallet_creation_exception.dart
