+++ oxen_coin/lib/src/exceptions/wallet_restore_from_keys_exception.dart
