+++ oxen_coin/lib/src/exceptions/wallet_restore_from_seed_exception.dart
