+++ oxen_coin/lib/src/native/account_list.dart
@@ -1,0 +1,38 @@
+import 'dart:ffi';
+
+import 'package:ffi/ffi.dart';
+import 'package:oxen_coin/src/oxen_api.dart';
+import 'package:oxen_coin/src/util/signatures.dart';
+import 'package:oxen_coin/src/util/types.dart';
+
+final accountSizeNative = oxenApi
+    .lookup<NativeFunction<account_size>>('account_size')
+    .asFunction<SubaddressSize>();
+
+final accountRefreshNative = oxenApi
+    .lookup<NativeFunction<account_refresh>>('account_refresh')
+    .asFunction<AccountRefresh>();
+
+final accountGetAllNative = oxenApi
+    .lookup<NativeFunction<account_get_all>>('account_get_all')
+    .asFunction<AccountGetAll>();
+
+final accountAddNewNative = oxenApi
+    .lookup<NativeFunction<account_add_new>>('account_add_row')
+    .asFunction<AccountAddNew>();
+
+final accountSetLabelNative = oxenApi
+    .lookup<NativeFunction<account_set_label>>('account_set_label_row')
+    .asFunction<AccountSetLabel>();
+
+void addAccountSync({String label}) {
+  final labelPointer = Utf8.toUtf8(label);
+  accountAddNewNative(labelPointer);
+  free(labelPointer);
+}
+
+void setLabelForAccountSync({int accountIndex, String label}) {
+  final labelPointer = Utf8.toUtf8(label);
+  accountSetLabelNative(accountIndex, labelPointer);
+  free(labelPointer);
+}
