+++ oxen_coin/lib/src/native/stake.dart
@@ -1,0 +1,49 @@
+import 'dart:ffi';
+
+import 'package:ffi/ffi.dart';
+import 'package:oxen_coin/oxen_coin_structs.dart';
+import 'package:oxen_coin/src/exceptions/creation_transaction_exception.dart';
+import 'package:oxen_coin/src/oxen_api.dart';
+import 'package:oxen_coin/src/structs/ut8_box.dart';
+import 'package:oxen_coin/src/util/signatures.dart';
+import 'package:oxen_coin/src/util/types.dart';
+
+final stakeCountNative = oxenApi
+    .lookup<NativeFunction<stake_count>>('stake_count')
+    .asFunction<StakeCount>();
+
+final stakeGetAllNative = oxenApi
+    .lookup<NativeFunction<stake_get_all>>('stake_get_all')
+    .asFunction<StakeGetAll>();
+
+final stakeCreateNative = oxenApi
+    .lookup<NativeFunction<stake_create>>('stake_create')
+    .asFunction<StakeCreate>();
+
+PendingTransactionDescription createStakeSync(String address, String amount) {
+  final addressPointer = Utf8.toUtf8(address);
+  final amountPointer = amount != null ? Utf8.toUtf8(amount) : nullptr;
+  final errorMessagePointer = allocate<Utf8Box>();
+  final pendingTransactionRawPointer = allocate<PendingTransactionRaw>();
+  final created = stakeCreateNative(addressPointer, amountPointer,
+          errorMessagePointer, pendingTransactionRawPointer) !=
+      0;
+
+  free(addressPointer);
+
+  if (amountPointer != nullptr) {
+    free(amountPointer);
+  }
+
+  if (!created) {
+    final message = errorMessagePointer.ref.getValue();
+    free(errorMessagePointer);
+    throw CreationTransactionException(message: message);
+  }
+
+  return PendingTransactionDescription(
+      amount: pendingTransactionRawPointer.ref.amount,
+      fee: pendingTransactionRawPointer.ref.fee,
+      hash: pendingTransactionRawPointer.ref.getHash(),
+      pointerAddress: pendingTransactionRawPointer.address);
+}
