+++ oxen_coin/lib/src/native/transaction_history.dart
@@ -1,0 +1,80 @@
+import 'dart:ffi';
+
+import 'package:ffi/ffi.dart';
+import 'package:oxen_coin/oxen_coin_structs.dart';
+import 'package:oxen_coin/src/exceptions/creation_transaction_exception.dart';
+import 'package:oxen_coin/src/oxen_api.dart';
+import 'package:oxen_coin/src/structs/ut8_box.dart';
+import 'package:oxen_coin/src/util/signatures.dart';
+import 'package:oxen_coin/src/util/types.dart';
+
+final transactionsRefreshNative = oxenApi
+    .lookup<NativeFunction<transactions_refresh>>('transactions_refresh')
+    .asFunction<TransactionsRefresh>();
+
+final transactionsCountNative = oxenApi
+    .lookup<NativeFunction<transactions_count>>('transactions_count')
+    .asFunction<TransactionsCount>();
+
+final transactionsGetAllNative = oxenApi
+    .lookup<NativeFunction<transactions_get_all>>('transactions_get_all')
+    .asFunction<TransactionsGetAll>();
+
+final transactionCreateNative = oxenApi
+    .lookup<NativeFunction<transaction_create>>('transaction_create')
+    .asFunction<TransactionCreate>();
+
+final transactionCommitNative = oxenApi
+    .lookup<NativeFunction<transaction_commit>>('transaction_commit')
+    .asFunction<TransactionCommit>();
+
+final transactionEstimateFeeNative = oxenApi
+    .lookup<NativeFunction<transaction_estimate_fee>>(
+        'transaction_estimate_fee')
+    .asFunction<TransactionEstimateFee>();
+
+PendingTransactionDescription createTransactionSync(
+    {String address, String amount, int priorityRaw, int accountIndex = 0}) {
+  final addressPointer = Utf8.toUtf8(address);
+  final amountPointer = amount != null ? Utf8.toUtf8(amount) : nullptr;
+  final errorMessagePointer = allocate<Utf8Box>();
+  final pendingTransactionRawPointer = allocate<PendingTransactionRaw>();
+  final created = transactionCreateNative(
+          addressPointer,
+          amountPointer,
+          priorityRaw,
+          accountIndex,
+          errorMessagePointer,
+          pendingTransactionRawPointer) !=
+      0;
+
+  free(addressPointer);
+
+  if (amountPointer != nullptr) {
+    free(amountPointer);
+  }
+
+  if (!created) {
+    final message = errorMessagePointer.ref.getValue();
+    free(errorMessagePointer);
+    throw CreationTransactionException(message: message);
+  }
+
+  return PendingTransactionDescription(
+      amount: pendingTransactionRawPointer.ref.amount,
+      fee: pendingTransactionRawPointer.ref.fee,
+      hash: pendingTransactionRawPointer.ref.getHash(),
+      pointerAddress: pendingTransactionRawPointer.address);
+}
+
+void commitTransaction({Pointer<PendingTransactionRaw> transactionPointer}) {
+  final errorMessagePointer = allocate<Utf8Box>();
+  final isCommited =
+      transactionCommitNative(transactionPointer, errorMessagePointer) != 0;
+
+  if (!isCommited) {
+    final message = errorMessagePointer.ref.getValue();
+    free(errorMessagePointer);
+    throw CreationTransactionException(message: message);
+  }
+}
