+++ oxen_coin/lib/src/structs/account_row.dart
