+++ oxen_coin/lib/src/structs/pending_transaction.dart
