+++ oxen_coin/lib/src/structs/stake_row.dart
