+++ oxen_coin/lib/src/structs/subaddress_row.dart
