+++ oxen_coin/lib/src/structs/transaction_info_row.dart
