+++ oxen_coin/lib/src/structs/ut8_box.dart
