+++ oxen_coin/lib/src/util/types.dart
@@ -1,8 +1,8 @@
 import 'dart:ffi';
 
 import 'package:ffi/ffi.dart';
-import 'package:oxen_coin/structs/pending_transaction.dart';
-import 'package:oxen_coin/structs/ut8_box.dart';
+import 'package:oxen_coin/oxen_coin_structs.dart';
+import 'package:oxen_coin/src/structs/ut8_box.dart';
 
 typedef CreateWallet = int Function(
     Pointer<Utf8>, Pointer<Utf8>, Pointer<Utf8>, int, Pointer<Utf8>);
